fix: remove mistakenly created components. Utilize what is already there
