--- typescript-react/src/components/ListingHistoryCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ListingHistoryCard';
