--- typescript-react/src/components/ListingHistoryCard/listing-history-card.module.scss
@@ -1,20 +1,0 @@
-@use '@styles/variables' as *;
-
-.listing-history-card {
-  width: 30%;
-  border: 1px solid $neutral-1;
-  border-radius: 4px;
-  margin-bottom: $space-m;
-  padding: $space-l;
-  background-color: $neutral-0;
-  box-shadow: $elevation-1;
-
-  &__price {
-    font-size: $font-l;
-  }
-
-  &__date {
-    color: $neutral-2;
-    font-style: italic;
-  }
-}
