+++ typescript-react/src/components/PriceHistoryCard/PriceHistoryCard.tsx
@@ -1,18 +1,29 @@
+import { FC } from 'react';
+
 import styles from './price-history-card.module.scss';
 
-const PriceHistoryCard = () => {
+import { PriceHistory } from '@/types/PriceHistoryTypes';
+import { convertTimestampToDate } from '@/utils/helpers';
+
+type PriceHistoryCardProps = {
+  items: PriceHistory[];
+};
+
+const PriceHistoryCard: FC<PriceHistoryCardProps> = ({ items }) => {
   return (
-    <div className={styles['container']}>
-      <table className={styles['price-card']}>
+    <div className={styles['price-history-card']}>
+      <table className={styles['price-history-card__price-card']}>
         <tbody>
-          <tr className={styles['price-card__header']}>
+          <tr className={styles['price-history-card__price-card__header']}>
             <th scope="col">Date</th>
             <th scope="col">Price (eur)</th>
           </tr>
-          <tr>
-            <td>01/01/2022</td>
-            <td>500 000</td>
-          </tr>
+          {items.map((item) => (
+            <tr>
+              <td>{convertTimestampToDate(item.created_date)}</td>
+              <td>{item.price_eur} &euro;</td>
+            </tr>
+          ))}
         </tbody>
       </table>
     </div>
