+++ typescript-react/src/components/PriceHistoryCard/price-history-card.module.scss
@@ -1,6 +1,6 @@
 @use '@styles/variables' as *;
 
-.container {
+.price-history-card {
   margin: $space-xl auto;
   align-items: baseline;
   border: 1px solid $neutral-1;
@@ -9,28 +9,28 @@
   padding: $space-l;
   background-color: $neutral-0;
   box-shadow: $elevation-1;
-}
 
-.price-card {
-  width: 100%;
-  align-items: baseline;
-  border: 1px solid $neutral-1;
-  border-radius: 4px;
-  grid-template-columns: 100%;
-  margin-bottom: $space-m;
-  padding: $space-l;
-}
+  &__price-card {
+    width: 100%;
+    align-items: baseline;
+    border: 1px solid $neutral-1;
+    border-radius: 4px;
+    grid-template-columns: 100%;
+    margin-bottom: $space-m;
+    padding: $space-l;
 
-.price-card__header {
-  border-bottom: 1px solid $neutral-1;
-}
+    &__header {
+      border-bottom: 1px solid $neutral-1;
+    }
 
-td,
-th {
-  border: 1px solid $neutral-1;
-  padding: 10px;
-}
+    td,
+    th {
+      border: 1px solid $neutral-1;
+      padding: 10px;
+    }
 
-td {
-  text-align: center;
+    td {
+      text-align: center;
+    }
+  }
 }
