--- typescript-react/src/containers/ListingHistory/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './ListingHistory';
