+++ typescript-react/src/containers/PricesHistory/PricesHistory.tsx
@@ -1,14 +1,28 @@
+import { useEffect, useState } from 'react';
 import PricesHistoryCard from '@components/PriceHistoryCard';
+import { useParams } from 'react-router-dom';
 
 import styles from './prices-history.module.scss';
 
-const PricesHistory = () => {
-  return (
-    <div className={styles['container']}>
-      <h1>Prices History</h1>
-      <PricesHistoryCard />
+import { PriceHistory } from '@/types/PriceHistoryTypes';
+import { getListingById } from '@/utils/requests';
 
-      <a href="/" className={styles['link']}>
+const PricesHistory = () => {
+  const { listingId } = useParams();
+  const [listingPrices, setListingPrices] = useState<PriceHistory[]>([]);
+
+  useEffect(() => {
+    getListingById(Number(listingId)).then((data) => {
+      setListingPrices(data);
+    });
+  }, [listingId]);
+
+  return (
+    <div className={styles['price-history']}>
+      <h1>Prices History</h1>
+      <PricesHistoryCard items={listingPrices} />
+
+      <a href="/" className={styles['price-history__link']}>
         &larr; Back Home
       </a>
     </div>
