+++ typescript-react/src/containers/PricesHistory/prices-history.module.scss
@@ -1,14 +1,14 @@
 @use '@styles/variables' as *;
 
-.container {
+.price-history {
   width: 80vw;
   margin: 0 auto;
-}
 
-.link {
-  border: solid 1px $primary-70;
-  padding: $space-s;
-  text-decoration: none;
-  color: $primary-70;
-  border-radius: 4px;
+  &__link {
+    border: solid 1px $primary-70;
+    padding: $space-s;
+    text-decoration: none;
+    color: $primary-70;
+    border-radius: 4px;
+  }
 }
