+++ typescript-react/src/types/PriceHistoryTypes.ts
@@ -1,4 +1,4 @@
-export type ListingPriceHistory = {
+export type PriceHistory = {
   created_date: string;
   price_eur: number;
 };
