+++ typescript-react/src/types/RequestsTypes.ts
@@ -1,9 +1,9 @@
 import { FormState } from './FormTypes';
-import { ListingPriceHistory } from './ListingsTypes';
+import { PriceHistory } from './PriceHistoryTypes';
 
 export type GetListingsResponse = Promise<FormState[]>;
 
-export type GetListingByIdResponse = Promise<ListingPriceHistory[]>;
+export type GetListingByIdResponse = Promise<PriceHistory[]>;
 
 export type CreateListingPayload = FormState;
 export type CreateListingResponse = Promise<FormState>;
